Prevent amp-story-page from being used anywhere other than root
